+++ smile_id_core/WebApi.py
@@ -64,13 +64,10 @@ class WebApi(Base):
         self,
         partner_params: Dict[str, Any],
         id_info_params: Dict[str, str],
-        use_validation_api: bool,
         options_params: OptionsParams,
     ) -> Dict[str, Any]:
         id_api = IdApi(self.partner_id, self.api_key, self.sid_server)
-        return id_api.submit_job(
-            partner_params, id_info_params, use_validation_api, options_params
-        )
+        return id_api.submit_job(partner_params, id_info_params, options_params)
 
     def submit_job(
         self,
@@ -78,7 +75,6 @@ class WebApi(Base):
         images_params: List[ImageParams],
         id_info_params: Dict[str, Any],
         options_params: OptionsParams,
-        use_validation_api: bool = True,
     ) -> Dict[str, Any]:
         """Perform key/parameter validation, creates zipped file and uploads."""
         Utilities.validate_partner_params(partner_params)
@@ -100,7 +96,6 @@ class WebApi(Base):
                     self.url,
                     id_info_params,
                     partner_params,
-                    use_validation_api,
                 )
                 id_info_params = {
                     "first_name": None,
@@ -129,7 +124,6 @@ class WebApi(Base):
             return self.__call_id_api(
                 partner_params,
                 id_info_params,
-                use_validation_api,
                 options_params,
             )
 
@@ -144,9 +138,7 @@ class WebApi(Base):
             use_enrolled_image=options_params.get("use_enrolled_image", False),
             job_type=job_type,
         )
-        Utilities.validate_id_params(
-            self.url, id_info_params, partner_params, use_validation_api
-        )
+        Utilities.validate_id_params(self.url, id_info_params, partner_params)
         self.__validate_return_data(options_params)
 
         signature_params = self.signature_params