+++ smile_id_core/constants.py
@@ -48,6 +48,9 @@ class JobType(IntEnum):
     # Compares document verification to an id check
     COMPARE_USER_INFO = 9
 
+    # Verifies user selfie with info retrieved from the ID issuing authority.
+    ENHANCED_DOCUMENT_VERIFICATION = 11
+
 
 # Defines production and sandbox test endpoints