+++ tests/test_id_api.py
@@ -81,7 +81,7 @@ def test_invalid_job_type(
     )
     partner_params["job_type"] = JobType.BIOMETRIC_KYC
     with pytest.raises(ValueError) as value_error:
-        client.submit_job(partner_params, id_info_params, False)
+        client.submit_job(partner_params, id_info_params, None)
     assert str(value_error.value) == "Job type must be 5 for ID Api"
 
 
@@ -168,81 +168,6 @@ def get_id_response(signature_fixture: Signature) -> Dict[str, Any]:
     }
 
 
-def get_smile_services_response() -> Dict[str, Any]:
-    """Returns supported Nigeria ID type examples for mocks"""
-    return {
-        "id_types": {
-            "NG": {
-                "NIN": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "job_id",
-                ],
-                "CAC": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "company",
-                    "job_id",
-                ],
-                "TIN": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "job_id",
-                ],
-                "VOTER_ID": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "job_id",
-                ],
-                "BVN": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "job_id",
-                ],
-                "PHONE_NUMBER": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "job_id",
-                    "first_name",
-                    "last_name",
-                ],
-                "DRIVERS_LICENSE": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "job_id",
-                    "first_name",
-                    "last_name",
-                    "dob",
-                ],
-                "PASSPORT": [
-                    "country",
-                    "id_type",
-                    "id_number",
-                    "user_id",
-                    "job_id",
-                    "first_name",
-                    "last_name",
-                    "dob",
-                ],
-            },
-        }
-    }
-
-
 def test_error_return_data(
     kyc_partner_params: Dict[str, Any],
     kyc_id_info: Dict[str, str],
@@ -270,12 +195,6 @@ def test_error_return_data(
 
             mocked_get.return_value.status_code = 200
             mocked_get.return_value.ok = True
-            mocked_get.return_value.text.return_value = (
-                get_smile_services_response()
-            )
-            mocked_get.return_value.json.return_value = (
-                get_smile_services_response()
-            )
 
             client.submit_job(partner_params, id_info_params)
     assert (
@@ -307,7 +226,7 @@ def test_validate_return_data(
             signature_fixture,
         )
 
-        response = client.submit_job(partner_params, id_info_params, False)
+        response = client.submit_job(partner_params, id_info_params, None)
 
         assert response is not None
 
@@ -320,6 +239,31 @@ def test_validate_return_data(
             response["IDNumber"]
             == get_id_response(signature_fixture)["IDNumber"]
         )
+
+
+def test_show_deprecation_warning_for_use_validation_api(
+    kyc_partner_params: Dict[str, Any],
+    kyc_id_info: Dict[str, str],
+    signature_fixture: Signature,
+    client: IdApi,
+) -> None:
+    """Uses mocked examples for payload responses; checks that
+    response status code is 200 and response body is not empty"""
+    partner_params, id_info_params = (
+        kyc_partner_params,
+        kyc_id_info,
+    )
+    with patch("requests.post") as mocked_post:
+        mocked_post.return_value.status_code = 200
+        mocked_post.return_value.ok = True
+        mocked_post.return_value.text.return_value = get_id_response(
+            signature_fixture,
+        )
+        mocked_post.return_value.json.return_value = get_id_response(
+            signature_fixture,
+        )
+
+        client.submit_job(partner_params, id_info_params)
 
 
 def test_validate_return_data_business_verification(
@@ -355,7 +299,7 @@ def test_validate_return_data_business_verification(
             signature_fixture,
         )
 
-        response = client.submit_job(partner_params, id_info_params, False)
+        response = client.submit_job(partner_params, id_info_params, None)
 
         assert response